pyproject.toml
[tool.poetry]
name = "shopping-assistant"
version = "0.1.0"
description = ""
authors = ["Daniele <lubrano93@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
lxml = "^4.9.3"
jinja2 = "^3.1.2"
pyyaml = "^6.0.1"
mailjet-rest = "^1.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
